requirements.txt
requests==2.28.0
boto3==1.9.4
toml==0.10.2
python-dotenv==0.20.0
sqlalchemy==1.4.39
mysql-connector-python==8.0.29
pandas==1.5.0